Import read_file_contents() from new script (to reduce duplication between pages)
