+++ pages/5_⚖️_License.py
@@ -1,39 +1,17 @@
 import streamlit as st
-import urllib.request as request
+from scripts.read_file import read_file_contents
 
 LICENSE_FILE = (
     "https://raw.githubusercontent.com/pythonhealthdatascience/"
     + "stars-streamlit-example/dev/LICENSE"
 )
 
-LICENSE_PAGE_TXT = (
-    "https://raw.githubusercontent.com/pythonhealthdatascience/"
-    + "stars-streamlit-example/dev/txt/license_page.md"
-)
+# LICENSE_PAGE_TXT = (
+#     "https://raw.githubusercontent.com/pythonhealthdatascience/"
+#     + "stars-streamlit-example/dev/txt/license_page.md"
+# )
 
-
-def read_file_contents(path):
-    """
-    Download the content of a file from the GitHub Repo and return as a utf-8 string
-
-    Notes:
-    -------
-        adapted from 'https://github.com/streamlit/demo-self-driving'
-
-    Parameters:
-    ----------
-    path: str
-        e.g. file_name.md
-
-    Returns:
-    --------
-    utf-8 str
-
-    """
-    response = request.urlopen(path)
-    return response.read().decode("utf-8")
-
-# show license header and badge
+# Show license header and badge
 #st.markdown(read_file_contents(LICENSE_PAGE_TXT))
 st.markdown("# License")
