+++ scripts/read_file.py
@@ -1,0 +1,23 @@
+import urllib.request as request
+
+
+def read_file_contents(path):
+    """
+    Download the content of a file from the GitHub Repo and return as a utf-8 string
+
+    Notes:
+    -------
+        adapted from 'https://github.com/streamlit/demo-self-driving'
+
+    Parameters:
+    ----------
+    path: str
+        e.g. file_name.md
+
+    Returns:
+    --------
+    utf-8 str
+
+    """
+    response = request.urlopen(path)
+    return response.read().decode("utf-8")
